third_party/blink/web_tests/svg/custom/a-focus-by-tab-expected.txt

tabToLinks=true
PASS input1.focus(); document.activeElement is input1
PASS eventSender.keyDown(tab); document.activeElement is document.getElementById("svg-a")
tabToLinks=false
PASS input1.focus(); document.activeElement is input1
PASS eventSender.keyDown(tab); document.activeElement is not document.getElementById("svg-a")
PASS successfullyParsed is true

TEST COMPLETE

 
Link
